refactor(codebuild-build): convert to libnuke resource format
